reorganize tracking qa
