+++ offline/packages/micromegas/Makefile.am
@@ -68,7 +68,6 @@ pkginclude_HEADERS = \
   CylinderGeomMicromegas.h \
   MicromegasCalibrationData.h \
   MicromegasClusterizer.h \
-  MicromegasClusterQA.h \
   MicromegasCombinedDataDecoder.h \
   MicromegasCombinedDataEvaluation.h \
   MicromegasDefs.h \
@@ -104,12 +103,10 @@ libmicromegas_io_la_LIBADD = \
   -lg4detectors_io \
   -lphool \
   -ltrack_io \
-  -ltrackbase_historic_io \
-  -lqautils
+  -ltrackbase_historic_io
 
 libmicromegas_la_SOURCES = \
   MicromegasClusterizer.cc \
-  MicromegasClusterQA.cc \
   MicromegasCombinedDataDecoder.cc \
   MicromegasCombinedDataEvaluation.cc \
   MicromegasCombinedDataEvaluation_Dict.cc \
